🧹 清理项目: 删除所有测试文件，只保留源码和文档
- 删除测试目录: tests/, test_files/, test-configs/, test-results/, test-logs/
- 删除日志文件: access.log, logs/
- 删除临时重构文档: REFACTOR_*.md
- 删除构建产物: anx可执行文件, build/目录
- 删除测试配置: server.conf
- 删除多余构建文件: Makefile.new, CMakeLists.txt
- 优化.gitignore，移除已删除文件的相关规则
- 保留核心源码、正式文档和必要配置文件

项目现在只包含生产就绪的代码和文档

+++ src/utils/asm/asm_integration.c
@@ -197,7 +197,7 @@ ssize_t asm_optimized_send(int socket_fd, const void* buffer, size_t len, int fl
             remaining -= sent;
             
             // 如果发送的数据少于请求的，可能是缓冲区满了
-            if (sent < chunk_size) {
+            if (sent < (ssize_t)chunk_size) {
                 break;
             }
         }
